Move all compat compressors into a single file
