+++ pandas/compat/__init__.py
@@ -12,9 +12,9 @@ from __future__ import annotations
 import os
 import platform
 import sys
-from typing import TYPE_CHECKING
 
 from pandas._typing import F
+import pandas.compat._compressors
 from pandas.compat.numpy import (
     is_numpy_dev,
     np_version_under1p21,
@@ -25,9 +25,6 @@ from pandas.compat.pyarrow import (
     pa_version_under8p0,
     pa_version_under9p0,
 )
-
-if TYPE_CHECKING:
-    import pandas.compat.lzma
 
 PY39 = sys.version_info >= (3, 9)
 PY310 = sys.version_info >= (3, 10)
@@ -121,7 +118,7 @@ def is_ci_environment() -> bool:
     return os.environ.get("PANDAS_CI", "0") == "1"
 
 
-def get_lzma_file() -> type[pandas.compat.lzma.LZMAFile]:
+def get_lzma_file() -> type[pandas.compat._compressors.LZMAFile]:
     """
     Importing the `LZMAFile` class from the `lzma` module.
 
@@ -135,15 +132,13 @@ def get_lzma_file() -> type[pandas.compat.lzma.LZMAFile]:
     RuntimeError
         If the `lzma` module was not imported correctly, or didn't exist.
     """
-    try:
-        import pandas.compat.lzma
-    except ImportError:
+    if not pandas.compat._compressors.has_lzma:
         raise RuntimeError(
             "lzma module not available. "
             "A Python re-install with the proper dependencies, "
             "might be required to solve this issue."
         )
-    return pandas.compat.lzma.LZMAFile
+    return pandas.compat._compressors.LZMAFile
 
 
 __all__ = [
