+++ pandas/compat/_compressors.py
@@ -1,0 +1,78 @@
+"""
+Patched ``BZ2File`` and ``LZMAFile`` to handle pickle protocol 5.
+"""
+
+from __future__ import annotations
+
+import bz2
+import pickle
+import sys
+
+try:
+    import lzma
+
+    has_lzma = True
+except ImportError:
+    has_lzma = False
+
+
+PY310 = sys.version_info >= (3, 10)
+
+
+def flatten_buffer(
+    b: bytes | bytearray | memoryview | pickle.PickleBuffer,
+) -> bytes | memoryview:
+    """
+    Return some 1-D `uint8` typed buffer.
+
+    Coerces anything that does not match that description to one that does
+    without copying if possible (otherwise will copy).
+    """
+
+    if isinstance(b, (bytes, bytearray)):
+        return b
+
+    if not isinstance(b, pickle.PickleBuffer):
+        b = pickle.PickleBuffer(b)
+
+    try:
+        # coerce to 1-D `uint8` C-contiguous `memoryview` zero-copy
+        return b.raw()
+    except BufferError:
+        # perform in-memory copy if buffer is not contiguous
+        return memoryview(b).tobytes("A")
+
+
+if not PY310:
+
+    class BZ2File(bz2.BZ2File):
+        def write(self, b) -> int:
+            # Workaround issue where `bz2.BZ2File` expects `len`
+            # to return the number of bytes in `b` by converting
+            # `b` into something that meets that constraint with
+            # minimal copying.
+            #
+            # Note: This is fixed in Python 3.10.
+            return super().write(flatten_buffer(b))
+
+    if has_lzma:
+
+        class LZMAFile(lzma.LZMAFile):
+            def write(self, b) -> int:
+                # Workaround issue where `lzma.LZMAFile` expects `len`
+                # to return the number of bytes in `b` by converting
+                # `b` into something that meets that constraint with
+                # minimal copying.
+                #
+                # Note: This is fixed in Python 3.10.
+                return super().write(flatten_buffer(b))
+
+else:
+
+    class BZ2File(bz2.BZ2File):
+        pass
+
+    if has_lzma:
+
+        class LZMAFile(lzma.LZMAFile):
+            pass
