--- pandas/compat/_utils.py
@@ -1,31 +1,0 @@
-"""
-Utilities used in various `compat` components.
-"""
-
-from __future__ import annotations
-
-import pickle
-
-
-def flatten_buffer(
-    b: bytes | bytearray | memoryview | pickle.PickleBuffer,
-) -> bytes | memoryview:
-    """
-    Return some 1-D `uint8` typed buffer.
-
-    Coerces anything that does not match that description to one that does
-    without copying if possible (otherwise will copy).
-    """
-
-    if isinstance(b, (bytes, bytearray)):
-        return b
-
-    if not isinstance(b, pickle.PickleBuffer):
-        b = pickle.PickleBuffer(b)
-
-    try:
-        # coerce to 1-D `uint8` C-contiguous `memoryview` zero-copy
-        return b.raw()
-    except BufferError:
-        # perform in-memory copy if buffer is not contiguous
-        return memoryview(b).tobytes("A")
